testing done

+++ controllers/favoriteissues.controller.js
@@ -2,20 +2,20 @@ const FAVORITEISSUES_MODEL = require("../models").Favorite_Issue;
 
 class FavoriteIssuesController {
   // POST New Favorite Issues
-	static postNewFavoriteIssues(req, res) {
-    try{
+  static postNewFavoriteIssues(req, res) {
+    try {
       const newFavoriteIssues = {
         user_id: req.userAccount.user_id,
-        issue_id: req.body.issue_id
-      }
+        issue_id: req.body.issue_id,
+      };
       FAVORITEISSUES_MODEL.create(newFavoriteIssues)
-        .then(result => {
-        res.status(200).json({ 
-            message: 'Success post new favorite issues!', 
-            result 
-         })
+        .then((result) => {
+          res.status(200).json({
+            message: "Success post new favorite issues!",
+            result,
+          });
         })
-        .catch(err => res.status(400).json({ message: err }))
+        .catch((err) => res.status(400).json({ message: err }));
     } catch (error) {
       res.status(500).send({
         error: error.message || "Internal Server Error",
@@ -48,18 +48,28 @@ class FavoriteIssuesController {
   // GET All Favorite Issue by User Id
   static async getFavoriteIssuesByUserId(req, res) {
     try {
-      const user = req.params.user
+      const user = req.params.id;
 
-      const dataFavIssues = await FAVORITEISSUES_MODEL.findAll({user : user})
+      const dataFavIssues = await FAVORITEISSUES_MODEL.findOne({
+        where: {
+          user_id: user,
+        },
+      });
 
       if (dataFavIssues) {
-        res.status(200).send({
-          message: `Success Get Your Favorite Issues`,
-          favoriteIssues: dataFavIssues,
-        });
+        if (Number(req.userAccount.user_id) === Number(user)) {
+          res.status(200).send({
+            message: "Success Get Your Favorite Issues",
+            favoriteIssues: dataFavIssues,
+          });
+        } else {
+          res.status(400).send({
+            message: "Data Favorite Issue is Not Yours",
+          });
+        }
       } else {
         res.status(404).send({
-          message: `Data Favorite Issue Not Found`,
+          message: "Data Favorite Issue is Not Found",
         });
       }
     } catch (error) {
@@ -69,38 +79,78 @@ class FavoriteIssuesController {
     }
   }
 
-  // DELETE Favorite Issue by Id
-  static async deleteFavoriteIssueById(req, res) {
+  // DELETE Favorite Issue by User Id
+  static async deleteFavoriteIssueByUserId(req, res) {
     try {
-      const issueID = req.params.id;
-      const userID = req.userAccount.user_id
-     
-      const dataFavIssues = await FAVORITEISSUES_MODEL.findOne({
+      const user = req.params.id;
+
+      const dataFavIssues = await FAVORITEISSUES_MODEL.findAll({
         where: {
-          issue_id: Number(issueID),
+          user_id: user,
         },
       });
 
       if (dataFavIssues) {
-        const favIssueUser = await FAVORITEISSUES_MODEL.destroy({
-          where: {
-            issue_id: Number(issueID),
-            user_id: Number(userID),
-          },
-        });
-        if(favIssueUser){
-          res.status(200).send({
-            message: `Data Favorite Issue where Issue Id is ${issueID} was Deleted Successfully`,
-            deletedFavoriteIssue: dataFavIssues,
+        if (Number(req.userAccount.user_id) === Number(user)) {
+          await FAVORITEISSUES_MODEL.destroy({
+            where: {
+              user_id: Number(user),
+            },
           });
-        } else{
-          res.status(404).send({
-            message: `Cannot delete Issue from Favorite because this Favorite Issue is not yours`,
+          res.status(200).send({
+            message: "Success Delete Your Favorite Issues",
+            favoriteIssues: dataFavIssues,
+          });
+        } else {
+          res.status(400).send({
+            message: "Data Favorite Issue is Not Yours",
           });
         }
       } else {
         res.status(404).send({
-          message: `Data Favorite Issue where Issue Id is ${issueID} Not Found`,
+          message: "Data Favorite Issue is Not Found",
+        });
+      }
+    } catch (error) {
+      res.status(500).send({
+        error: error.message || "Internal Server Error",
+      });
+    }
+  }
+
+  // DELETE Favorite Issue by Issue Id
+  static async deleteFavoriteIssueByIssueId(req, res) {
+    try {
+      const issue = req.params.issueId;
+      const user = req.params.userId;
+
+      const dataFavIssues = await FAVORITEISSUES_MODEL.findOne({
+        where: {
+          user_id: user,
+          issue_id: issue,
+        },
+      });
+
+      if (dataFavIssues) {
+        if (Number(req.userAccount.user_id) === Number(user)) {
+          await FAVORITEISSUES_MODEL.destroy({
+            where: {
+              issue_id: Number(issue),
+              user_id: Number(user),
+            },
+          });
+          res.status(200).send({
+            message: "Success Delete Your Favorite Issues",
+            favoriteIssues: dataFavIssues,
+          });
+        } else {
+          res.status(400).send({
+            message: "Data Favorite Issue is Not Yours",
+          });
+        }
+      } else {
+        res.status(404).send({
+          message: "Data Favorite Issue is Not Found",
         });
       }
     } catch (error) {
